pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "popup-controller"
version = "0.1.0"
description = "PySide6 Windows desktop application for controlling and flashing an ESP32 pop-up controller over serial."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-2.0-or-later" }
authors = [
  { name = "Sheep" },
]
classifiers = [
  "License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "PySide6>=6.7,<7.0",
  "esptool>=5.2,<6.0",
  "pyserial>=3.5,<4.0",
]

[project.optional-dependencies]
dev = [
  "pyinstaller>=6.0,<7.0",
  "pytest>=8.0,<9.0",
  "pytest-qt>=4.4,<5.0",
  "ruff>=0.6,<1.0",
]

[project.scripts]
popup-controller = "popup_controller.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/popup_controller"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
